% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{corrp_pal}
\alias{corrp_pal}
\title{Return function to interpolate a CORRP colour palette}
\usage{
corrp_pal(palette = "Main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of the registry (e.g. bcca, excel, acfdr, etc)}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a CORRP colour palette
}
\keyword{internal}
