% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_adnet_report}
\alias{create_adnet_report}
\title{Create ReadMe.md files}
\usage{
create_adnet_report(
  report = c("validation", "invoice", "internal"),
  registry_token = getPass(msg = "API token for REDCap project: ADNeT Registry"),
  holding_token = getPass(msg = "API token for REDCap project: ADNeT Holding Database"),
  optout = getPass(msg = "API token for REDCap project: ADNeT Opt-out Database")
)
}
\arguments{
\item{report}{Which report should be created (options are 'validation', 'invoice', 'internal')}

\item{registry_token}{API token for ADNET registry}

\item{holding_token}{API token for ADNeT Holding Database}

\item{optout}{API token for ADNeT Opt-out Database}
}
\value{
This function creates the desired readme files. It will not overwrite the file
however if it does not exists. It does not return anything.
}
\description{
This function creates ADNET reports
file structure
}
\keyword{ReadMe}
\keyword{ReadMe.md}
