% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_readme.R
\name{create_readme}
\alias{create_readme}
\title{Create ReadMe.md files}
\usage{
create_readme(
  template = c("assets", "rmds", "raw_data", "produced_data", "plots", "tables",
    "reports", "functions", "rscripts", "adhoc_req"),
  path = getwd()
)
}
\arguments{
\item{template}{Which ReadMe.md files should be created? Default is all. Partial
matching does work.}

\item{path}{Where should they be created? Default is the working directory.}
}
\value{
This function creates the desired readme files. It will not overwrite the file
however if it does not exists. It does not return anything.
}
\description{
This function creates ReadMe.md in the CORRP templates for the standard
file structure
}
\keyword{ReadMe}
\keyword{ReadMe.md}
