% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_data}
\alias{fake_data}
\title{Status of 40,000 patients}
\format{
A data frame with 40000 rows and 9 variables:
\describe{
  \item{patient_id}{patint unique id}
  \item{site_id}{site unique id}
  ...
}
}
\usage{
fake_data
}
\description{
A dataset containing the demographics and status of 40,000
patinets from 20 sites.
}
\keyword{datasets}
