% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.R
\name{fplot}
\alias{fplot}
\title{Make funnel plots}
\usage{
fplot(data, outcome, num, denom, bm, ci1 = 0.95, ci2 = 0.998)
}
\arguments{
\item{data}{dataframe containing the numertaors and denominators}

\item{outcome}{string health outcome (e.g. "inpatient mortality")}

\item{num}{the name of the column in the dataframe which includes the numerator (e.g. numerator)}

\item{denom}{the name of the column in the dataframe which includes the denominator (e.g. denominator)}

\item{bm}{the benchamrk value}

\item{ci1}{the alpha related to the first line of control limits (e.g. 0.95)}

\item{ci2}{the alpha related to the second line of control limits (e.g 0.998)}
}
\value{
ggplot
}
\description{
Create a funnel plot from a dataframe containing at least two columns (i.e numerator and denominator)
It is currently working only with binary outcome, continuous outcomes will be supported in future
}
\examples{

}
