% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Identifies outliers using a robust z-score}
\usage{
outlier(x, n = 3)
}
\arguments{
\item{x}{the numeric vector of interest}

\item{n}{the number of median absolute deviations from the median outside which to assign a datum as an outlier (default = 3)}
}
\value{
the outlier status (1, 0)
}
\description{
This function takes one argument, the number of median absolute deviations outside which to assign a particular data point as an outlier.
}
