% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_pyramid.R
\name{population_pyramid}
\alias{population_pyramid}
\title{Create a population pyramid plot}
\usage{
population_pyramid(data)
}
\arguments{
\item{data}{dataframe which must include three variables (sex, age_group, and n)}
}
\value{
pyramid plot as a ggplot2 object
}
\description{
This function assumes three variables in the dataset: 1. sex (i.e. 'Male' vs 'Female)
2. age_group 3. n (count)
}
\examples{

}
