% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Produce reports}
\usage{
render_report(
  rmd,
  n_file,
  open_file = TRUE,
  report_dir = "reports",
  rmds_dir = "rmds"
)
}
\arguments{
\item{rmd}{name of the rmd file}

\item{n_file}{prefix numbering of the output file}

\item{open_file}{open the html file in the default browser once produced. By default, \code{TRUE}.}

\item{report_dir}{target directory for the reports}

\item{rmds_dir}{source directery for the \code{.Rmds} files. By default, the directory \emph{rmds}.}
}
\value{
Return invisibly the file path of the produced report
}
\description{
Produce html reports from a rmd file. It's mainly a wrapper arrond \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
