% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_adjust.R
\name{risk_adjust}
\alias{risk_adjust}
\title{Make funnel plots}
\usage{
risk_adjust(data, site_id, outcome, outcome_type, adj_vars)
}
\arguments{
\item{data}{dataframe containing the numertaors and denominators}

\item{site_id}{string health service (e.g. "site_code")}

\item{outcome}{string health outcome (e.g. "inpatient mortality")}

\item{outcome_type}{type of the outcome variables (i.e. "binary" or "continuous")}

\item{adj_vars}{vector with the adjusting variables (e.g. c("sex", "age"))}
}
\value{
the risk adjusted dataset
}
\description{
Create a funnel plot from a dataframe containing at least two columns (i.e numerator and denominator)
It is currently working only with binary outcome, continuous outcomes will be supported in future
}
\examples{

}
