% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_color_corrp_discrete}
\alias{scale_color_corrp_discrete}
\title{Discrete color scale based on CORRP colors}
\usage{
scale_color_corrp_discrete(palette, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in corrp_palettes (e.g. "bcca", "excel", "acfdr", etc)}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
Discrete color scale based on CORRP colors
}
