% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_fill_corrp_continuous}
\alias{scale_fill_corrp_continuous}
\title{Continuous fill scale based on corrp colors}
\usage{
scale_fill_corrp_continuous(palette, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in corrp_palettes ("Main" or "Blues")}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to scale_color_gradientn()}
}
\description{
Continuous fill scale based on corrp colors
}
