% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_fill_corrp_discrete}
\alias{scale_fill_corrp_discrete}
\title{Discrete fill scale based on corrp colors}
\usage{
scale_fill_corrp_discrete(palette, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in corrp_palettes ("Main" or "Blues")}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
Discrete fill scale based on corrp colors
}
