% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_funnelcount.R
\name{stat_funnelcount}
\alias{stat_funnelcount}
\title{Add funnel plot boundaries.}
\usage{
stat_funnelcount(
  mapping = NULL,
  data = NULL,
  geom = "ribbon",
  group = ci_cat,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ci_limits = 0.95,
  step = 0.5,
  ...
)
}
\arguments{
\item{ci_limits}{The limit to be used in computing the boundary
(default = \code{0.95}).}

\item{step}{Step to use in drawing the boundary
(default = \code{0.5}).}
}
\value{
ggplot stat
}
\description{
Add funnel plot boundaries.
}
