% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{stepwise}
\alias{stepwise}
\title{Perform a stepwise logistic or linear regression using LRT tests of significance.}
\usage{
stepwise(
  out_var,
  ind_var,
  pe = 0.01,
  pr = 0.05,
  data = NULL,
  model = "logistic"
)
}
\arguments{
\item{out_var}{the outcome variable}

\item{ind_var}{the independent variables to consider}

\item{pe}{the significance level above which a variable may enter the model}

\item{pr}{the significance level below which a variable may be deleted from the model}

\item{data}{the data frame to use}

\item{model}{the model that needs to be performed ("logistic" or "linear")}
}
\value{
the final model
}
\description{
Copied and edited from: https://msu.edu/~rubin/code/stepwise_demo.nb.html
}
