% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableby_to_kable.R
\name{tableby_to_kable}
\alias{tableby_to_kable}
\title{Convert an arsenal::tableby object to a kable}
\usage{
tableby_to_kable(tableby, label_translation, caption, ref_label)
}
\arguments{
\item{tableby}{tableby object}

\item{label_translation}{A named list (or vector) where the name is the label in the output to be replaced in the pretty rendering by the character string value for the named element of the list, e.g., list(age = "Age(Years)", meansd = "Mean(SD)")}

\item{caption}{Title/caption for the table, defaulting to NULL (no title).}

\item{ref_label}{The table reference label}
}
\value{
a kable object
}
\description{
Create a funnel plot from a dataframe containing at least two columns (i.e numerator and denominator)
It is currently working only with binary outcome, continuous outcomes will be supported in future
}
\examples{

}
