% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.R
\name{fplot}
\alias{fplot}
\title{Make funnel plots}
\usage{
fplot(
  data,
  outcome,
  num,
  denom,
  bm,
  ci1 = 0.95,
  ci2 = 0.998,
  cols = c("#F8766D", "#00BFC4", "#7CAE00"),
  site_name
)
}
\arguments{
\item{data}{dataframe containing the numertaors and denominators}

\item{outcome}{string health outcome (e.g. "inpatient mortality")}

\item{num}{the name of the column in the dataframe which includes the numerator (e.g. numerator)}

\item{denom}{the name of the column in the dataframe which includes the denominator (e.g. denominator)}

\item{bm}{the benchamrk value}

\item{ci1}{the alpha related to the first line of control limits (e.g. 0.95)}

\item{ci2}{the alpha related to the second line of control limits (e.g 0.998)}

\item{cols}{the colours used for the points above, within and below the control first control limits}

\item{site_name}{the sites to be highlighted}
}
\value{
ggplot
}
\description{
Create a funnel plot from a dataframe containing at least two columns (i.e numerator and denominator)
It is currently working only with binary outcome, continuous outcomes will be supported in future
}
\examples{

}
