% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Identifies outliers using a robust z-score}
\usage{
outlier(data, n = 3)
}
\arguments{
\item{data}{the data frame for which outliers need to be found}

\item{n}{the number of median absolute deviations from the median outside which to assign a datum as an outlier (default = 3)}
}
\value{
the dataframe with outlier status (1, 0) for all the numerical variables
}
\description{
This function takes a dataframe and one argument (i.e. the number of median absolute deviations outside which to assign a particular data point as an outlier).
}
\examples{

}
